requirements.txt
fastapi
uvicorn
websockets
python-multipart
pdfplumber
google-generativeai
edge-tts
pydantic
gTTS
python-dotenv